+++ instrumentation/opentelemetry-instrumentation-sqlalchemy/src/opentelemetry/instrumentation/sqlalchemy/__init__.py
@@ -104,7 +104,7 @@ class SQLAlchemyInstrumentor(BaseInstrumentor):
 
         if kwargs.get("engine") is not None:
             return EngineTracer(
-                _get_tracer(kwargs.get("engine"), tracer_provider),
+                _get_tracer(tracer_provider),
                 kwargs.get("engine"),
                 kwargs.get("enable_commenter", False),
             )